优化库存同步 更改库存金额字段类型

See merge request bms/scmp/scmp-api!350

+++ src/main/java/com/aiqin/bms/scmp/api/purchase/service/impl/PurchaseApplyServiceImpl.java
@@ -369,7 +369,7 @@ public class PurchaseApplyServiceImpl implements PurchaseApplyService {
                 if(info != null){
                     detail.setStockCount(info.getInventoryNum().intValue());
                     detail.setTotalWayNum(info.getTotalWayNum().intValue());
-                    detail.setStockAmount(info.getInventoryNum().intValue() * info.getTaxCost().intValue());
+                    detail.setStockAmount(info.getInventoryNum().longValue() * info.getTaxCost().longValue());
                     detail.setNewPurchasePrice(info.getNewPurchasePrice().intValue());
                 }
             }
